requirements.txt
undetected-chromedriver==3.5.5
selenium==4.21.0
pymongo==4.9.2
python-dotenv==1.0.1
setuptools>=68.0.0
requests==2.32.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2